requirements.txt
# Core dependencies
streamlit>=1.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.3.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.3

# Document processing
unstructured>=0.10.0
pypdf>=3.0.0
pandas>=2.0.0
openpyxl>=3.0.0

# Embeddings & Vector Storage
chromadb>=0.4.0
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0

# File handling
python-magic>=0.4.27
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0
nltk>=3.8.0

# Optional formats
jsonpath-ng>=1.5.0